removed two player support code
fixed the reverting conflicts
